+++ src/constants.js
@@ -29,9 +29,9 @@ export const PAGE_SCRIPT_ELEMENT_NAME = 'skip-to-content';
 export const BOOKMARKLET_ELEMENT_NAME = 'skip-to-content-bookmarklet';
 export const EXTENSION_ELEMENT_NAME   = 'skip-to-content-extension';
 
-export const INFO_DIALOG_ELEMENT_NAME = 'skip-to-content-info-dialog-575';
-export const MESSAGE_ELEMENT_NAME     = 'skip-to-content-message-element-575';
-export const HIGHLIGHT_ELEMENT_NAME   = 'skip-to-content-highlight-element-575';
+export const INFO_DIALOG_ELEMENT_NAME = 'skip-to-content-info-dialog-576';
+export const MESSAGE_ELEMENT_NAME     = 'skip-to-content-message-element-576';
+export const HIGHLIGHT_ELEMENT_NAME   = 'skip-to-content-highlight-element-576';
 
 // Attributes